+++ xrpl/transaction/vault_withdraw.go
@@ -92,7 +92,7 @@ func (tx *VaultWithdraw) Validate() (bool, error) {
 		}
 	}
 
-	if tx.CredentialIDs != nil && !tx.CredentialIDs.IsValidForWithdrawal() {
+	if tx.CredentialIDs != nil && !tx.CredentialIDs.IsValid() {
 		return false, ErrInvalidCredentialIDs
 	}